fix memory leak bug in c

+++ shortest_path_opt/include/dp_info.h
@@ -27,11 +27,11 @@
 typedef struct _DPInfo
 {
     struct _DPInfo *parent;
-    int latest_insert_idx;
+    long latest_insert_idx;
     double cost;
 
-    int (*getSequenceLength)(struct _DPInfo *self);
-    int* (*getSequence)(struct _DPInfo *self);
+    long (*getSequenceLength)(struct _DPInfo *self);
+    long* (*getSequence)(struct _DPInfo *self);
     void (*setParent)(struct _DPInfo *self, struct _DPInfo *parent);
 }DPInfo;
 
@@ -40,14 +40,14 @@ typedef struct _DPInfoObject{
     DPInfo *dp_info;
 } DPInfoObject;
 
-int getDPInfoSequenceLength(DPInfo *self);
-int *getDPInfoSequence(DPInfo *self);
+long getDPInfoSequenceLength(DPInfo *self);
+long *getDPInfoSequence(DPInfo *self);
 void setDPInfoParent(DPInfo *self, DPInfo *parent);
 
-void initDPInfo(DPInfo *self, int latest_insert_idx);
+void initDPInfo(DPInfo *self, long latest_insert_idx);
 void freeDPInfo(DPInfo *self);
 
-DPInfo* newDPInfo(int latest_insert_idx);
+DPInfo* newDPInfo(long latest_insert_idx);
 
 extern PyTypeObject DPInfoType;
